Allow to override entrypoint with other commands.
The current entry point doesn't allow to override the location
of the configuration, making it very hard to use any other
command to run gnatsd with.

This change allows users to start the container with an clean
command giving them access to change all configuration options,
beside the location of the configuration file.

Signed-off-by: David Calavera <david.calavera@gmail.com>