Reorganize layout
This change reorganizes the folder layout to allow us to write 1 image
definition for several architectures. This means we can avoid multiple
arch-Directory declarations in the official nats IMF file.

In addition, this should make it easier to add additional image bases in
the future, for example, if we wanted to add an alpine version of nats.

Finally, this adds GitHub Actions CI to make sure that the images build.