Tweak image strategy
Based on feedback from docker-library/official-images, we're reverting
the use of multi-stage builds because they don't work as you'd expect
in docker-library/official-images.

In this change, we'll use the CGO enabled official release binaries for
all containers that support CGO. For images that don't support CGO, such
as scratch, we'll use special locally built binaries in this repo with
CGO disabled, like before.