Support multiple versions
The directory structure is based on minor series, e.g. 2.9.x.
Given an update from 2.9.22 to 2.9.23 (for example), the
existing directory will be used, but the files updated in place.

If however, the update is for 2.10.0, it will find the latest minor
version and create a new minor directory based off that.

Signed-off-by: Byron Ruth <byron@nats.io>