Fix and scope tests per version
The tests attempted to handle both versions and ended up
being buggy when performing an version update due to drift
in the alpine version, for instance.

The tests have been scoped to each version directory and simplified.
The update script now only operates on one of version directories
at a time.

Signed-off-by: Byron Ruth <byron@nats.io>